API: change default behavior of shifts in FT
The new behavior is to default to shifts that ensure the zero
frequency to lie in the grid. Users may still override this,
but no longer for half-complex transforms. Here, only the default
setting is valid since FFT shift should be used.

TODO: implement FFT shifts in the FT operators